Trouble with a few power up tools on RemNote
every time I type "##" and initial the tag "extra card detail" it does not appear to exist, nor does it function. The only success I have is to copy the tag link then pasting it to the rem. It has become very annoying as I used to be able to just be able to use that power up without difficulty